Refactoring Application to implement CQRS.
