+++ EShop.Product.Api/EShop.Product.Api.csproj
@@ -18,6 +18,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\EShop.Infrastructure\EShop.Infrastructure.csproj" />
+    <ProjectReference Include="..\EShop.Product.DataProvider\EShop.Product.DataProvider.csproj" />
   </ItemGroup>
 
 </Project>
