+++ EShop.Product.DataProvider/Repository/IProductRepository.cs
@@ -1,11 +1,11 @@
 ﻿using EShop.Infrastructure.Command.Product;
 using EShop.Infrastructure.Event.Product;
 
-namespace EShop.Product.Api.Service
+namespace EShop.Product.DataProvider.Repository
 {
-    public interface IProductService
+    public interface IProductRepository
     {
-        Task<ProductCreated> GetProduct(string ProductId);
         Task<ProductCreated> AddProduct(CreateProduct product);
+        Task<ProductCreated> GetProduct(string ProductId);
     }
 }
